% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check_packages}
\alias{check_packages}
\title{Check if packages are required and install if not.}
\usage{
check_packages(package)
}
\arguments{
\item{package}{List of packages to be installed.}
}
\description{
Utility function.
}
