% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctt_norms.R
\name{ctt_norms}
\alias{ctt_norms}
\title{Computes Colour Trails Test (CTT) norms.}
\usage{
ctt_norms(
  education,
  age,
  male = TRUE,
  source = "Lee2012_Singapore",
  trails1 = NA,
  trails2 = NA
)
}
\arguments{
\item{education}{Education in years.}

\item{age}{Age in years.}

\item{male}{TRUE or FALSE.}

\item{source}{Can be `Lee2012_Singapore` or ` Hsieh2007_China`.}

\item{trails1}{Trails 1 performance in seconds.}

\item{trails2}{Trails 2 performance in seconds.}
}
\value{
A table of standardised scores and their descriptors. Note that z-scores were computed using `(mean - x) / sd` formula, hence larger z-scores correspond to better performance.
}
\description{
Computes Colour Trails Test (CTT) norms.
}
\examples{
out <- ctt_norms(education=10, age=55, trails1=30, trails2=40, source="Hsieh2007_China");
}
