% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvlt_norms.R
\name{cvlt_norms}
\alias{cvlt_norms}
\title{Computes California Verbal Learning Test (CVLT) norms.}
\usage{
cvlt_norms(
  education,
  age,
  male = TRUE,
  source = "Whittle2007_US",
  version = "II",
  form = "short",
  t1 = NA,
  t2 = NA,
  t3 = NA,
  t4 = NA,
  t5 = NA,
  listB = NA,
  t6 = NA,
  total = NA,
  delayed_free_recall = NA,
  delayed_cued_recall = NA,
  recog_correct = NA,
  recog_FA = NA
)
}
\arguments{
\item{education}{Education in years.}

\item{age}{Age in years.}

\item{male}{TRUE or FALSE.}

\item{source}{Only computes `Whittle2007_US` norms (CVLT-II short form) for now.}

\item{version}{Character string. Only computes `II` for now (i.e., CVLT-II)}

\item{form}{Character string. Only computes `short` for now (i.e., CVLT short form)}

\item{t1}{Trial 1 score}

\item{t2}{Trial 2 score}

\item{t3}{Trial 3 score}

\item{t4}{Trial 4 score}

\item{t5}{Trial 5 score}

\item{listB}{List B score}

\item{t6}{Trial 6 score}

\item{total}{Totals (Trials 1-5 for standard form; Trails 1-4 for short form)}

\item{delayed_free_recall}{Delayed Free Recall score}

\item{delayed_cued_recall}{Delayed Cued Recall score}

\item{recog_correct}{Recognition total number of correct items/hits.}

\item{recog_FA}{Recognition False Alarms.}
}
\value{
A table of standardised scores and their descriptors.
}
\description{
Computes California Verbal Learning Test (CVLT) norms.
}
\examples{
out1 <- cvlt_norms(education=10, age=80, t1=5, t2=5, t3=6, t4=6, total=22, delayed_free_recall=10);
}
