% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravlt_norms.R
\name{ravlt_norms}
\alias{ravlt_norms}
\title{Computes Rey Auditory Verbal Learning Test (RAVLT) norms.}
\usage{
ravlt_norms(
  education,
  age,
  male = TRUE,
  source = "Lee2012_Singapore",
  t1 = NA,
  t2 = NA,
  t3 = NA,
  t4 = NA,
  t5 = NA,
  listB = NA,
  t6 = NA,
  total = NA,
  delayed_recall = NA,
  recog_correct = NA,
  recog_FA = NA
)
}
\arguments{
\item{education}{Education in years.}

\item{age}{Age in years.}

\item{male}{TRUE or FALSE.}

\item{source}{Can be `Lee2012_Singapore` or ` Wallace2017_ChineseAus`.}

\item{t1}{Trial 1 score}

\item{t2}{Trial 2 score}

\item{t3}{Trial 3 score}

\item{t4}{Trial 4 score}

\item{t5}{Trial 5 score}

\item{listB}{List B score}

\item{t6}{Trial 6 score (Short delay recall)}

\item{total}{Totals (Trials 1-5) score.}

\item{delayed_recall}{Delayed Free Recall score}

\item{recog_correct}{Recognition total number of correct items/hits.}

\item{recog_FA}{Recognition False Alarms.}
}
\value{
A table of standardised scores and their descriptors.
}
\description{
Computes Rey Auditory Verbal Learning Test (RAVLT) norms.
}
\examples{
out1 <- ravlt_norms(education=10, age=56, t1=5, t2=4, t3=6, t4=6, t5=10, total=15, listB=2, t6=6, delayed_recall=10, recog_correct=10, recog_FA=2);
}
