% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmt_norms.R
\name{sdmt_norms}
\alias{sdmt_norms}
\title{Computes Symbol Digit Modality Test (SDMT) norms.}
\usage{
sdmt_norms(
  education,
  age,
  male = TRUE,
  source = "Kiely2014_Aus_Written",
  written = NA,
  verbal = NA
)
}
\arguments{
\item{education}{Education in years.}

\item{age}{Age in years.}

\item{male}{TRUE or FALSE.}

\item{source}{Can be `Kiely2014_Aus_Written` or ` Hsieh2007_China`.}

\item{written}{Written performance score.}

\item{verbal}{Verbal performance score.}
}
\value{
A table of standardised scores and their descriptors.
}
\description{
Computes Symbol Digit Modality Test (SDMT) norms.
}
\examples{
out1 <- sdmt_norms(education="postsecondary", age=20, male=TRUE, written=50, verbal=50);
out2 <-  sdmt_norms(education="tertiary", age=60, male=TRUE, written=50, verbal=50);
}
