% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt_norms.R
\name{tmt_norms}
\alias{tmt_norms}
\title{Computes Trail Making Test (TMT) norms.}
\usage{
tmt_norms(
  education,
  age,
  male = TRUE,
  source = "Tombaugh2004_Canada",
  trailsA = NA,
  trailsB = NA
)
}
\arguments{
\item{education}{Education in years.}

\item{age}{Age in years.}

\item{male}{TRUE or FALSE.}

\item{source}{Can be `Tombaugh2004_Canada`, `Hsieh2007_China`, or ` Whittle2007_US`.}

\item{trailsA}{Trails A performance in seconds.}

\item{trailsB}{Trails B performance in seconds.}
}
\value{
A table of standardised scores and their descriptors. Note that z-scores were computed using `(mean - x) / sd` formula, hence larger z-scores correspond to better performance.
}
\description{
Computes Trail Making Test (TMT) norms.
}
\examples{
out1 <- tmt_norms(education=10, age=55, trailsA=30, trailsB=40);
out2 <- tmt_norms(education=10, age=91, trailsA=30, trailsB=40, source="Whittle2007_US")
}
